Use int not float to store ints
These are now real bugs found by enabling these warnings.

+++ src/Filters/ZJetCutsFilter.h
@@ -43,7 +43,7 @@ class MinNMuonsCut : public ZJetFilterBase
 	}
 
   private:
-	float nMuonsMin;
+	unsigned int nMuonsMin;
 };
 
 
@@ -70,7 +70,7 @@ class MaxNMuonsCut : public ZJetFilterBase
 	}
 
   private:
-	float nMuonsMax;
+	unsigned int nMuonsMax;
 };
